separate MigrationData.cs
