+++ Assets/VRM10/Runtime/IO/MigrationData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86d0f8ee9681a2e43b959952e51810cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
